Split admin insights formatting helpers
