+++ tests/admin-insights-architecture.test.ts
@@ -4,12 +4,16 @@ import test from 'node:test';
 
 const pagePath = 'frontend/app/(core)/admin/insights/page.tsx';
 const helpersPath = 'frontend/app/(core)/admin/insights/_lib/insights-helpers.ts';
+const formattersPath = 'frontend/app/(core)/admin/insights/_lib/insights-formatters.ts';
+const navigationPath = 'frontend/app/(core)/admin/insights/_lib/insights-navigation.ts';
 const typesPath = 'frontend/app/(core)/admin/insights/_lib/insights-types.ts';
 const panelsPath = 'frontend/app/(core)/admin/insights/_components/InsightsPanels.tsx';
 
 test('admin insights page stays a route orchestrator', () => {
   assert.equal(existsSync(pagePath), true);
   assert.equal(existsSync(helpersPath), true);
+  assert.equal(existsSync(formattersPath), true);
+  assert.equal(existsSync(navigationPath), true);
   assert.equal(existsSync(typesPath), true);
   assert.equal(existsSync(panelsPath), true);
 
@@ -19,6 +23,7 @@ test('admin insights page stays a route orchestrator', () => {
   assert.ok(pageLines < 260, `expected admin insights page to stay under 260 lines, got ${pageLines}`);
   assert.match(pageSource, /from '\.\/_lib\/insights-types';/);
   assert.match(pageSource, /from '\.\/_lib\/insights-helpers';/);
+  assert.match(pageSource, /from '\.\/_lib\/insights-navigation';/);
   assert.match(pageSource, /from '\.\/_components\/InsightsPanels';/);
   assert.match(pageSource, /fetchAdminMetrics\(/);
   assert.match(pageSource, /fetchAdminMetricsComparison\(/);
@@ -30,15 +35,33 @@ test('admin insights page stays a route orchestrator', () => {
   assert.doesNotMatch(pageSource, /new Intl\./);
 });
 
-test('admin insights helpers own data shaping and formatting', () => {
+test('admin insights helpers own data shaping', () => {
   const helpersSource = readFileSync(helpersPath, 'utf8');
 
   assert.match(helpersSource, /export function buildExecutiveMetrics/);
   assert.match(helpersSource, /export function buildFocusMetricData/);
   assert.match(helpersSource, /export function buildRevenueBoardRows/);
   assert.match(helpersSource, /export function buildMonthlyRows/);
-  assert.match(helpersSource, /export function describeRange/);
-  assert.match(helpersSource, /const currencyFormatter = new Intl\.NumberFormat/);
+  assert.match(helpersSource, /from '\.\/insights-formatters';/);
+  assert.match(helpersSource, /from '\.\/insights-navigation';/);
+  assert.doesNotMatch(helpersSource, /const currencyFormatter = new Intl\.NumberFormat/);
+  assert.doesNotMatch(helpersSource, /export function buildInsightsHref/);
+  assert.doesNotMatch(helpersSource, /export function resolveFocusParam/);
+});
+
+test('admin insights formatting and navigation helpers stay split', () => {
+  const formattersSource = readFileSync(formattersPath, 'utf8');
+  const navigationSource = readFileSync(navigationPath, 'utf8');
+
+  assert.match(formattersSource, /const currencyFormatter = new Intl\.NumberFormat/);
+  assert.match(formattersSource, /export function formatCurrency/);
+  assert.match(formattersSource, /export function formatCompactNumber/);
+  assert.match(formattersSource, /export function toneBadgeClass/);
+  assert.match(formattersSource, /export function resolveDeltaTone/);
+  assert.match(navigationSource, /export const FOCUS_OPTIONS/);
+  assert.match(navigationSource, /export function buildInsightsHref/);
+  assert.match(navigationSource, /export function resolveFocusParam/);
+  assert.match(navigationSource, /export function describeRange/);
 });
 
 test('admin insights panels own route-local JSX surfaces', () => {
@@ -50,4 +73,6 @@ test('admin insights panels own route-local JSX surfaces', () => {
   assert.match(panelsSource, /export function HealthPanel/);
   assert.match(panelsSource, /export function InsightsControls/);
   assert.match(panelsSource, /export function DailyLedgerTable/);
+  assert.match(panelsSource, /from '\.\.\/_lib\/insights-formatters';/);
+  assert.match(panelsSource, /from '\.\.\/_lib\/insights-navigation';/);
 });
